[REFACTOR] introduce AttributeEntityTable
