+++ frontend/src/pages/Attributes/ViewModal.vue
@@ -8,60 +8,7 @@
     >
       <template #default>
         <h3 class="q-my-md">{{ t('entity.basics') }}</h3>
-        <EntityViewTable>
-          <EntityViewTableRow :label="t('entity.commonColumns.name')">
-            {{ attribute.name }}
-          </EntityViewTableRow>
-          <EntityViewTableRow :label="t('attributes.columns.dataType')">
-            {{ dataTypeToLabel(attribute.data_type, t) }}
-          </EntityViewTableRow>
-          <EntityViewTableRow :label="t('attributes.minLong')">
-            {{ attribute.validation_rule?.min }}
-          </EntityViewTableRow>
-          <EntityViewTableRow :label="t('attributes.maxLong')">
-            {{ attribute.validation_rule?.max }}
-          </EntityViewTableRow>
-          <EntityViewTableRow :label="t('attributes.step')">
-            {{ attribute.validation_rule?.step }}
-          </EntityViewTableRow>
-          <EntityViewTableRow
-            :label="t('attributes.columns.defaultValue')"
-            multiline
-          >
-            <template v-if="defaultValue !== ''">
-              {{ defaultValue }}
-            </template>
-            <span v-else class="text-body2 text-italic">{{
-              t('base.notAvailable')
-            }}</span>
-          </EntityViewTableRow>
-          <template v-if="attribute.data_type === 'RATING' && attribute.legend">
-            <EntityViewTableRow
-              v-for="(label, idx) in attribute.legend"
-              :key="idx"
-              :label="
-                t('attributes.columns.legend') +
-                ` ${idx + attribute.validation_rule.min}`
-              "
-            >
-              {{ label }}
-            </EntityViewTableRow>
-          </template>
-          <EntityViewTableRow
-            v-if="attribute.description"
-            :label="t('attributes.columns.description')"
-            multiline
-          >
-            {{ attribute.description }}
-          </EntityViewTableRow>
-          <EntityViewTableRow :label="t('attributes.columns.attributeType')">
-            {{ attributeTypeToLabel(attribute.attribute_type, t) }}
-          </EntityViewTableRow>
-          <EntityTableViewTimestampRows
-            :created="attribute.created"
-            :modified="attribute.modified"
-          />
-        </EntityViewTable>
+        <AttributeEntityTable :attribute="attribute" />
 
         <h3 class="q-my-md">{{ t('attributes.preview') }}</h3>
         <AttributePreview :attribute="attribute" />
@@ -91,18 +38,9 @@ import type { AttributeFragment } from 'src/components/Attribute/attributeFragme
 import { attributeFragment } from 'src/components/Attribute/attributeFragment';
 import { useI18n } from 'src/composables/useI18n';
 import { useRoute, useRouter } from 'vue-router';
-import EntityViewTable from 'src/components/Entity/View/EntityViewTable.vue';
-import EntityViewTableRow from 'src/components/Entity/View/EntityViewTableRow.vue';
-import {
-  dataTypeToLabel,
-  attributeTypeToLabel,
-  dataTypeToColumnTypes,
-  formatResultColumnValue,
-} from 'src/utils/attributeUtils';
-import { ColumnTypes } from 'src/utils/columnTypes';
 import AttributePreview from 'src/components/Attribute/AttributePreview.vue';
-import EntityTableViewTimestampRows from 'src/components/Entity/View/EntityViewTableTimestampRows.vue';
 import EntityFetchWrapper from 'src/components/Entity/EntityFetchWrapper.vue';
+import AttributeEntityTable from 'src/components/Attribute/AttributeEntityTable.vue';
 
 const props = defineProps<{ entityId: number | string }>();
 
@@ -127,22 +65,7 @@ const attribute = computed(
   () => (data.value?.attributes_by_pk || null) as AttributeFragment | null,
 );
 
-const { t, d, n } = useI18n();
-
-const defaultValue = computed(() => {
-  const type =
-    attribute.value && dataTypeToColumnTypes(attribute.value?.data_type);
-
-  if (!type || type === ColumnTypes.Photo) {
-    return '';
-  }
-  return formatResultColumnValue({
-    value: attribute.value?.default_value,
-    type,
-    d,
-    n,
-  });
-});
+const { t } = useI18n();
 
 const route = useRoute();
 const router = useRouter();
